+++ .github/workflows/dotnet-build-pullrequest.yml
@@ -23,7 +23,7 @@ jobs:
     - name: Setup .NET Core
       uses: actions/setup-dotnet@v5
       with:
-        dotnet-version: '9.0.x' # Specify your .NET Core version, e.g., 6.x, 7.x
+        dotnet-version: '10.0.x' # Match TargetFramework net10.0 in uTPro.Project.Web.csproj
 
     - name: Restore dependencies
       run: dotnet restore ./uTPro/Project/uTPro.Project.Web/uTPro.Project.Web.csproj
