+++ .github/workflows/dotnet-release.yml
@@ -34,10 +34,10 @@ jobs:
           PREFIX=$(echo "$BRANCH" | cut -d'/' -f1)
           echo "prefix=$PREFIX" >> $GITHUB_OUTPUT
 
-      - name: Setup .NET 9 SDK
+      - name: Setup .NET 10 SDK
         uses: actions/setup-dotnet@v5
         with:
-          dotnet-version: '9.0.x'
+          dotnet-version: '10.0.x'
 
       - name: Build & Publish .NET project
         run: |
