+++ uTPro/Extension/uTPro.Extension.CurrentSite/DICurrentItemExtension.cs
@@ -1,0 +1,11 @@
+using Microsoft.Extensions.DependencyInjection;
+using Umbraco.Cms.Core.Composing;
+
+namespace uTPro.Extension.CurrentSite
+{
+    class DICurrentItemExtension : IComposer
+    {
+        public void Compose(IUmbracoBuilder builder)
+            => builder.Services.AddScoped<ICurrentItemExtension, CurrentItemExtension>();
+    }
+}
