+++ uTPro/Extension/uTPro.Extension/HttpContextStaticComposer.cs
@@ -1,4 +1,3 @@
-using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
@@ -20,24 +19,6 @@ namespace uTPro.Extension
 
             // Runs after the service provider is built to capture the accessor.
             builder.Services.AddSingleton<IStartupFilter, HttpContextStaticStartupFilter>();
-        }
-    }
-
-    internal sealed class HttpContextStaticStartupFilter : IStartupFilter
-    {
-        private readonly IHttpContextAccessor _accessor;
-
-        public HttpContextStaticStartupFilter(IHttpContextAccessor accessor)
-        {
-            _accessor = accessor;
-            HttpContextStatic.Accessor = accessor;
-        }
-
-        public Action<IApplicationBuilder> Configure(Action<IApplicationBuilder> next)
-        {
-            // Re-assign defensively in case something reset it. Cheap no-op otherwise.
-            HttpContextStatic.Accessor = _accessor;
-            return next;
         }
     }
 }
