+++ uTPro/Foundation/uTPro.Foundation.Favicon/IFaviconFoundation.cs
@@ -1,0 +1,7 @@
+namespace uTPro.Foundation.Favicon
+{
+    public interface IFaviconFoundation
+    {
+        string Generate();
+    }
+}
