+++ uTPro/Foundation/uTPro.Foundation.Middleware/RequestLocalizationOptionMiddleware.cs
@@ -11,22 +11,6 @@ using uTPro.Extension.CurrentSite;
 
 namespace uTPro.Foundation.Middleware
 {
-    public static class UseWebRequestLocalizationMiddleware
-    {
-        public static IApplicationBuilder UseWebRequestLocalization(this IApplicationBuilder app)
-        {
-            var config = app.ApplicationServices.GetRequiredService<IConfiguration>();
-            bool.TryParse(config.GetSection(ConfigSettingUTPro.ListRememberLanguage.Enabled)?.Value, out bool isEnabled);
-            if (isEnabled)
-            {
-                var requestLocalizationOptions = app.ApplicationServices.GetRequiredService<IOptions<RequestLocalizationOptions>>();
-                app.UseRequestLocalization(requestLocalizationOptions.Value);
-                return app.UseMiddleware<RequestLocalizationOptionMiddleware>();
-            }
-            return app;
-        }
-    }
-
     internal class RequestLocalizationOptionMiddleware
     {
         private const string CookieCulture = ".uTPro.Culture";
