+++ uTPro/Foundation/uTPro.Foundation.Robots/IRobotsFoundation.cs
@@ -1,0 +1,7 @@
+namespace uTPro.Foundation.Robots
+{
+    public interface IRobotsFoundation
+    {
+        string Generate();
+    }
+}
