+++ uTPro/Foundation/uTPro.Foundation.Sitemap/ISitemapFoundation.cs
@@ -1,0 +1,7 @@
+namespace uTPro.Foundation.Sitemap
+{
+    public interface ISitemapFoundation
+    {
+        string Generate();
+    }
+}
