+++ uTPro/Project/uTPro.Project.Web.Configure/DIICheckPolicy.cs
@@ -1,0 +1,11 @@
+using Microsoft.Extensions.DependencyInjection;
+using Umbraco.Cms.Core.Composing;
+
+namespace uTPro.Project.Web.Configure
+{
+    class DIICheckPolicy : IComposer
+    {
+        public void Compose(IUmbracoBuilder builder)
+            => builder.Services.AddScoped<ICheckPolicy, CheckPolicy>();
+    }
+}
