+++ uTPro/Project/uTPro.Project.Web.Configure/ICheckPolicy.cs
@@ -1,0 +1,9 @@
+using Microsoft.AspNetCore.Http;
+
+namespace uTPro.Project.Web.Configure
+{
+    public interface ICheckPolicy
+    {
+        string Check(HttpContext httpContext);
+    }
+}
