+++ uTPro/Project/uTPro.Project.Web/uTPro.Project.Web.csproj
@@ -3,7 +3,8 @@
 		<TargetFramework>net10.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
-		<CompressionEnabled>false</CompressionEnabled>
+		<!-- SDK build/publish-time Brotli+GZip pre-compression of static web assets (.NET 9+). -->
+		<CompressionEnabled>true</CompressionEnabled>
     <Version>17.1.0</Version>
   </PropertyGroup>
 	<ItemGroup>
